refactor: replace custom JoinHandle wrapper with From<JoinError> impl
Remove the custom `JoinHandle` wrapper that only existed to convert
`tokio::task::JoinError` to `ErrorCode`. Use idiomatic Rust error
conversion via the `From` trait instead, allowing the `?` operator
to handle conversion automatically.

Changes:
- Add `impl From<tokio::task::JoinError> for ErrorCode` with panic message extraction
- Remove custom `JoinHandle` struct from `runtime.rs`
- Re-export `tokio::task::JoinHandle` directly from runtime module
- Update call sites from `.await.flatten()` to `.await??` or `.await?`
