+++ src/common/base/src/runtime/mod.rs
@@ -53,7 +53,6 @@ pub use perf::QueryPerf;
 pub use perf::QueryPerfGuard;
 pub use runtime::Dropper;
 pub use runtime::GLOBAL_TASK;
-pub use runtime::JoinHandle;
 pub use runtime::Runtime;
 pub use runtime::block_on;
 pub use runtime::execute_futures_in_parallel;
@@ -79,3 +78,4 @@ pub use time_series::TimeSeriesProfileName;
 pub use time_series::TimeSeriesProfiles;
 pub use time_series::compress_time_point;
 pub use time_series::get_time_series_profile_desc;
+pub use tokio::task::JoinHandle;
