+++ src/common/exception/Cargo.toml
@@ -36,6 +36,7 @@ serde_json = { workspace = true }
 sqlx = { workspace = true }
 tantivy = { workspace = true }
 thiserror = { workspace = true }
+tokio = { workspace = true }
 tonic = { workspace = true }
 
 [lints]
